backend/requirements.txt
numpy==1.25.2
openai_whisper==20240930
opencv_python==4.11.0.86
openvino==2023.3.0
openvino_dev==2023.3.0
PyAudio==0.2.14
pydantic==2.11.7
torch==2.7.1
torchaudio==2.7.1
pyttsx3==2.90
TTS==0.17.6
flask==2.0.1
flask-cors==3.0.10
